# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "首页"
                ],
                "summary": "首页",
                "responses": {
                    "200": {
                        "description": "code\",\"msg\",\"data\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/conf/add": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "es配置"
                ],
                "summary": "添加es配置",
                "parameters": [
                    {
                        "description": "es配置",
                        "name": "newConfig",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.ESConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"msg\",\"data\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/conf/del": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "es配置"
                ],
                "summary": "删除es配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "索引名称",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"msg\",\"data\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/conf/get": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "es配置"
                ],
                "summary": "获取es配置",
                "responses": {
                    "200": {
                        "description": "code\",\"msg\",\"data\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/conf/use": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "es配置"
                ],
                "summary": "设置es配置",
                "parameters": [
                    {
                        "description": "es配置",
                        "name": "newConfig",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.ESConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"msg\",\"data\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/es/getIndices": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "es查询"
                ],
                "summary": "获取es索引",
                "responses": {
                    "200": {
                        "description": "code\",\"msg\",\"data\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/es/getMappings": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "es查询"
                ],
                "summary": "获取es索引mappings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "索引名称",
                        "name": "index",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"msg\",\"data\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ESConfig": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "selected": {
                    "type": "boolean"
                },
                "username": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        }
    }
}